chore: move monorepo-definitions out of the module
