+++ lib/monorepo.js
@@ -1,8 +1,9 @@
 const dbs = require('../lib/dbs')
+const { monorepoDefinitions } = require('../utils/monorepo-definitions')
 
 function getMonorepoGroup (name) {
-  return Object.keys(module.exports.monorepoDefinitions).find(group =>
-    module.exports.monorepoDefinitions[group].find(pkg => pkg === name))
+  return Object.keys(monorepoDefinitions).find(group =>
+    monorepoDefinitions[group].find(pkg => pkg === name))
 }
 
 function isPartOfMonorepo (dependency) {
@@ -16,7 +17,7 @@ async function hasAllMonorepoUdates (dependency, version) {
     // read all registry changes
     const docs = await npm.allDocs({
       include_docs: true,
-      keys: module.exports.monorepoDefinitions[group]
+      keys: monorepoDefinitions[group]
     })
 
     const docsWithoutErrors = docs.rows
@@ -33,7 +34,7 @@ async function hasAllMonorepoUdates (dependency, version) {
     }, {})
 
     // check if we have other monorepo packages
-    const receivedAllregistryChanges = module.exports.monorepoDefinitions[group].reduce((acc, packageName) => {
+    const receivedAllregistryChanges = monorepoDefinitions[group].reduce((acc, packageName) => {
       const pkg = registryChanges[packageName]
       // this is the main bit
       if (pkg && pkg.distTags['latest'] && pkg.distTags['latest'] !== version) {
@@ -47,57 +48,8 @@ async function hasAllMonorepoUdates (dependency, version) {
   }
 }
 
-// const monorepoDefinitions = {
-//  "prefix": [
-//    "prefix-full-name",
-//    "prefix-full-other"
-//  ]
-// }
-const monorepoDefinitions = {
-  'pouchdb': [
-    'pouchdb',
-    'pouchdb-abstract-mapreduce',
-    'pouchdb-adapter-fruitdown',
-    'pouchdb-adapter-http',
-    'pouchdb-adapter-idb',
-    'pouchdb-adapter-indexeddb',
-    'pouchdb-adapter-leveldb',
-    'pouchdb-adapter-leveldb-core',
-    'pouchdb-adapter-localstorage',
-    'pouchdb-adapter-memory',
-    'pouchdb-adapter-node-websql',
-    'pouchdb-adapter-utils',
-    'pouchdb-adapter-websql',
-    'pouchdb-adapter-websql-core',
-    'pouchdb-binary-utils',
-    'pouchdb-browser',
-    'pouchdb-changes-filter',
-    'pouchdb-checkpointer',
-    'pouchdb-collate',
-    'pouchdb-collections',
-    'pouchdb-core',
-    'pouchdb-debug',
-    'pouchdb-errors',
-    'pouchdb-fetch',
-    'pouchdb-find',
-    'pouchdb-for-coverage',
-    'pouchdb-generate-replication-id',
-    'pouchdb-json',
-    'pouchdb-mapreduce',
-    'pouchdb-mapreduce-utils',
-    'pouchdb-md5',
-    'pouchdb-merge',
-    'pouchdb-node',
-    'pouchdb-replication',
-    'pouchdb-selector-core',
-    'pouchdb-utils',
-    'sublevel-pouchdb'
-  ]
-}
-
 module.exports = {
   isPartOfMonorepo,
   hasAllMonorepoUdates,
-  getMonorepoGroup,
-  monorepoDefinitions
+  getMonorepoGroup
 }
