+++ utils/monorepo-definitions.js
@@ -1,0 +1,56 @@
+// const monorepoDefinitions = {
+//  "prefix": [
+//    "prefix-full-name",
+//    "prefix-full-other"
+//  ]
+// }
+
+const monorepoDefinitions = {
+  'pouchdb': [
+    'pouchdb',
+    'pouchdb-abstract-mapreduce',
+    'pouchdb-adapter-fruitdown',
+    'pouchdb-adapter-http',
+    'pouchdb-adapter-idb',
+    'pouchdb-adapter-indexeddb',
+    'pouchdb-adapter-leveldb',
+    'pouchdb-adapter-leveldb-core',
+    'pouchdb-adapter-localstorage',
+    'pouchdb-adapter-memory',
+    'pouchdb-adapter-node-websql',
+    'pouchdb-adapter-utils',
+    'pouchdb-adapter-websql',
+    'pouchdb-adapter-websql-core',
+    'pouchdb-binary-utils',
+    'pouchdb-browser',
+    'pouchdb-changes-filter',
+    'pouchdb-checkpointer',
+    'pouchdb-collate',
+    'pouchdb-collections',
+    'pouchdb-core',
+    'pouchdb-debug',
+    'pouchdb-errors',
+    'pouchdb-fetch',
+    'pouchdb-find',
+    'pouchdb-for-coverage',
+    'pouchdb-generate-replication-id',
+    'pouchdb-json',
+    'pouchdb-mapreduce',
+    'pouchdb-mapreduce-utils',
+    'pouchdb-md5',
+    'pouchdb-merge',
+    'pouchdb-node',
+    'pouchdb-replication',
+    'pouchdb-selector-core',
+    'pouchdb-utils',
+    'sublevel-pouchdb'
+  ]
+}
+
+// function getMonorepoDefinitions () {
+//
+// }
+
+module.exports = {
+  monorepoDefinitions
+}
